Rename middleware.py to utils.py
