+++ policy_wizard/views.py
@@ -9,7 +9,7 @@ from django.views.decorators.csrf import csrf_exempt
 from django.core.exceptions import ObjectDoesNotExist
 from .forms import NewPolicyForm
 from .models import PolicyTemplates, Policies
-from .middleware import role_identifier, validate_request
+from .utils import role_identifier, validate_request
 from .forms import PolicyTemplateForm, NewPolicyForm
 from django.views.decorators.clickjacking import xframe_options_exempt
